+++ src/pages/viewer/Channels.vue
@@ -34,7 +34,10 @@
               <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">see deleted</q-btn>
             </q-popover>
           </q-btn>
-          <div v-if="!items.length">{{isLoading ? 'Fetching data..' : 'Channels not found'}}</div>
+          <div v-if="!items.length">
+            <div>{{isLoading ? 'Fetching data..' : 'Channels not found'}}</div>
+            <q-btn v-if="!isLoading && needShowGetDeletedAction && tokenType === 1" class="q-mt-sm" @click="getDeletedHandler" icon="mdi-download" label="see deleted"/>
+          </div>
         </div>
       </div>
     </template>
@@ -140,7 +143,6 @@
         v-if="+size[1]"
         :style="{minHeight: `calc(${size[1]}vh - ${+size[0] ? isVisibleToolbar ? '50px' : '25px' : isVisibleToolbar ? '100px' : '50px'})`, position: 'relative'}"
       />
-      <div class="text-center" style="font-size: 1.5rem; margin-top: 30px; color: white" v-if="selectedItem.deleted">Nothing to show by channel &#171;{{selectedItem.name}}&#187; <div style="font-size: 0.9rem">or you haven`t access</div></div>
     </template>
   </q-page>
 </template>