+++ src/pages/viewer/Devices.vue
@@ -32,7 +32,10 @@
             <q-btn icon="mdi-download" class="deleted-action" @click="getDeletedHandler" v-if="needShowGetDeletedAction && tokenType === 1">see deleted</q-btn>
           </q-popover>
         </q-btn>
-        <div v-if="!items.length">{{isLoading ? 'Fetching data..' : 'Devices not found'}}</div>
+        <div v-if="!items.length">
+          <div>{{isLoading ? 'Fetching data..' : 'Devices not found'}}</div>
+          <q-btn v-if="!isLoading && needShowGetDeletedAction && tokenType === 1" class="q-mt-sm" @click="getDeletedHandler" icon="mdi-download" label="see deleted"/>
+        </div>
       </div>
     </div>
     <template v-else>
@@ -146,7 +149,6 @@
         @map:close="isVisibleMap = false"
         @map:minimize="mapMinimizeHandler"
       />
-      <div class="text-center" style="font-size: 1.5rem; margin-top: 30px; color: white" v-if="selectedItem.deleted">Nothing to show by device &#171;{{selectedItem.name || `#${selectedItem.id}`}}&#187; <div style="font-size: 0.9rem">or you haven`t access</div></div>
     </template>
   </q-page>
 </template>